Start refactoring models from norm to dataset + threat model
