+++ model_info/cifar10/L2/Augustin2020Adversarial.json
