+++ model_info/cifar10/L2/Rice2020Overfitting.json
