+++ model_info/cifar10/L2/Rony2019Decoupling.json
