+++ model_info/cifar10/L2/Standard.json
