+++ model_info/cifar10/L2/Wu2020Adversarial.json
