+++ model_info/cifar10/Linf/Alfarra2020ClusTR.json
