+++ model_info/cifar10/Linf/Atzmon2019Controlling.json
