+++ model_info/cifar10/Linf/Chan2020Jacobian.json
