+++ model_info/cifar10/Linf/Chen2020Adversarial.json
