+++ model_info/cifar10/Linf/Huang2020Self.json
