+++ model_info/cifar10/Linf/Jang2019Adversarial.json
