+++ model_info/cifar10/Linf/JinRinard2020Manifold.json
