+++ model_info/cifar10/Linf/Pang2020Rethinking.json
