+++ model_info/cifar10/Linf/Rice2020Overfitting.json
