+++ model_info/cifar10/Linf/Sehwag2020Hydra.json
