+++ model_info/cifar10/Linf/Shafahi2019Adversarial.json
