+++ model_info/cifar10/Linf/Standard.json
