+++ model_info/cifar10/Linf/Wong2020Fast.json
