+++ model_info/cifar10/Linf/Wu2020Adversarial.json
