+++ robustbench/model_zoo/architectures/resnext/METADATA
@@ -1,0 +1,12 @@
+name: "ResNeXt-DenseNet"
+description: "PyTorch implementations of ResNeXt and DenseNet."
+
+third_party {
+  url {
+    type: GIT
+    value: "https://github.com/D-X-Y/ResNeXt-DenseNet"
+  }
+  version: "0de9a8c8fd095b37eb60945f8dafefdbfe1cef6b"
+  last_upgrade_date { year: 2019 month: 12 day: 4 }
+  license_type: PERMISSIVE
+}
