+++ robustbench/model_zoo/enums.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+
+class BenchmarkDataset(Enum):
+    cifar_10 = 'cifar10'
+    cifar_100 = 'cifar_100'
+    image_net = 'imagenet'
+
+
+class ThreatModel(Enum):
+    L2 = "L2"
+    Linf = "Linf"
+    corruptions = "corruptions"
